Refactors rest API into `rest.rs` and `rest_utils.rs`
